test[notask]: cover NodeNext declaration specifiers
